requirements.txt
alembic==0.8.7
aniso8601==1.1.0
blinker==1.4
click==6.6
factory-boy==2.7.0
fake-factory==0.6.0
Flask==0.11.1
flask-marshmallow==0.7.0
Flask-Migrate==2.0.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.5.0
hiredis==0.2.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.9.1
marshmallow-enum==1.0
marshmallow-sqlalchemy==0.10.0
psycopg2==2.6.2
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.6.1
six==1.10.0
SQLAlchemy==1.0.14
SQLAlchemy-Utils==0.32.9
webargs==1.3.4
Werkzeug==0.11.10
